+++ lib/src/schema/schemas/comment_schema.dart
@@ -1,0 +1,14 @@
+class CommentSchema {
+  static const id =
+      'https://peiffer-innovations.github.io/flutter_json_schemas/schemas/json_dynamic_widget/comment.json';
+
+  static final schema = {
+    r'$schema': 'http://json-schema.org/draft-06/schema#',
+    r'$id': '$id',
+    r'$comment':
+        'https://pub.dev/documentation/json_dynamic_widget/latest/json_dynamic_widget/JsonCommentBuilder-class.html',
+    'type': 'object',
+    'title': 'Comment',
+    'additionalProperties': true,
+  };
+}