+++ test/builders/json_sliver_padding_builder_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:json_dynamic_widget/json_dynamic_widget.dart';
+
+void main() {
+  test('type', () {
+    const type = JsonSliverPaddingBuilder.type;
+
+    expect(type, 'sliver_padding');
+    expect(
+      JsonWidgetRegistry.instance.getWidgetBuilder(type)({})
+          is JsonSliverPaddingBuilder,
+      true,
+    );
+  });
+}